Remove GraphQL client dependency
