+++ Gemfile
@@ -38,9 +38,6 @@ gem 'bootsnap', '>= 1.4.4', require: false
 gem 'devise', github: 'heartcombo/devise'
 gem 'omniauth-github', '~> 1.1'
 
-# Connect to GitHub’s API
-gem 'graphql-client'
-
 gem 'autoprefixer-rails'
 gem 'font-awesome-sass'
 gem 'simple_form'
