+++ app/entities/github/api.rb
@@ -1,0 +1,66 @@
+class Github::Api
+  def each_commit(username, repository, cursor = nil, &block)
+    variables = { username: username, repository: repository, cursor: cursor }
+    data = query(COMMITS_BY_REPOSITORY_QUERY, variables: variables)
+    edges = data.dig(
+      "repository",
+      "defaultBranchRef",
+      "target",
+      "history",
+      "edges",
+    )
+    return if !edges || edges.count == 0
+
+    edges.each { |edge| block.call(edge["node"]) }
+
+    each_commit(username, repository, edges.last["cursor"], &block)
+
+    sleep 3
+  end
+
+  private
+
+  GRAPHQL_URL = "https://api.github.com/graphql"
+
+  COMMITS_BY_REPOSITORY_QUERY = <<~QUERY
+    query($username: String!, $repository: String!, $cursor: String) {
+      repository(owner: $username, name: $repository) {
+        defaultBranchRef {
+          target {
+            ... on Commit {
+              history(first: 100, after: $cursor) {
+                edges {
+                  cursor
+                  node {
+                    id
+                    message
+                    committedDate
+                    author {
+                      user {
+                        login
+                      }
+                    }
+                    committer {
+                      email
+                    }
+                  }
+                }
+              }
+            }
+          }
+        }
+      }
+    }
+  QUERY
+
+  def query(query, variables: {})
+    RestClient.post(
+      GRAPHQL_URL,
+      { query: query, variables: variables }.to_json,
+      "Authorization" => "Bearer #{ENV["GITHUB_TOKEN"]}",
+      "Content-Type" => "application/json"
+    ) do |response, _request, _result|
+      JSON.parse(response.body, symbolize_names: true)
+    end
+  end
+end
